==> 10_label_transfer_predictions.ipynb <==
import pandas as pd
import numpy as np
from tqdm.notebook import tqdm
from label_transfer_predictions import LabelTransferPredictor
import seaborn as sns
import matplotlib.pyplot as plt
from math import sqrt

datasets = pd.read_csv('datasets.csv')
datasets = datasets[~datasets['study_id'].isin(['Ho2020M', 'Ho2020F'])]

predictor = LabelTransferPredictor(hyperparams_file='hyperparams.csv',
                               data_dir='exported_matrices',
                               output_dir='label_transfer_model_output/predictions/',
                               verbose=False,
                               random_seed=12345)

for dataset in datasets['study_id'].unique():    
    print(f'Predicting labels for {dataset}')
    predictions = predictor.predict_dataset(dataset)
    predictor.plot_predictions_on_UMAP(dataset, predictions, display=False, save=True)  
    # Save predictions to file
    predictions.to_csv(f'label_transfer_model_output/{dataset}_label_transfer_predictions.csv', index=False)
